# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: 2PUC Notes Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sections/{id}/cards:
    get:
      operationId: listSectionCards
      tags: [sections]
      summary: List visible cards in a section
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
  /cards:
    get:
      operationId: listCards
      tags: [cards]
      summary: List/search/filter cards
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: subject
          in: query
          schema:
            type: string
        - name: resourceType
          in: query
          schema:
            type: string
        - name: isFree
          in: query
          schema:
            type: boolean
        - name: sectionId
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
  /cards/{id}:
    get:
      operationId: getCard
      tags: [cards]
      summary: Get a card by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Card detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "404":
          description: Not found
...
  /checkout/create-razorpay-order:
    post:
      operationId: createRazorpayOrder
      tags: [checkout]
      summary: Create a Razorpay order for a paid card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RazorpayOrderInput"
      responses:
        "200":
          description: Razorpay order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RazorpayOrderResult"
        "400":
          description: Bad request
  /checkout/verify:
    post:
      operationId: verifyCheckout
      tags: [checkout]
      summary: Verify Razorpay payment signature and mark order paid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCheckoutInput"
      responses:
        "200":
          description: Order confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid signature
  /checkout/free-download:
    post:
      operationId: freeDownload
      tags: [checkout]
      summary: Create a free-download order and return a download URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FreeDownloadInput"
      responses:
        "200":
          description: Download URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreeDownloadResult"
        "400":
          description: Bad request

  # ── Purchases ──────────────────────────────────────────────────────────────
  /purchases:
    get:
      operationId: getPurchases
      tags: [purchases]
      summary: Get purchases by phone number
      parameters:
        - name: phone
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /purchases/{orderId}/download/{cardId}:
    get:
      operationId: downloadPdf
      tags: [purchases]
      summary: Redirect to a 1-hour signed Supabase download URL
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "302":
          description: Redirect to signed URL
        "403":
          description: Not authorized

  # ── Admin ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Card:
      type: object
      properties:
        id:
          type: integer
        sectionId:
          type: integer
        sectionName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        chapter:
          type: ["string", "null"]
        semester:
          type: ["string", "null"]
        board:
          type: string
        classLevel:
          type: string
        resourceType:
          type: string
        isFree:
          type: boolean
        pricePaise:
          type: integer
        discountPricePaise:
          type: ["integer", "null"]
        telegramLink:
          type: ["string", "null"]
        isNew:
          type: boolean
        thumbnailUrl:
          type: ["string", "null"]
        previewImageUrls:
          type: array
          items:
            type: string
        pdfFileKey:
          type: ["string", "null"]
        pageCount:
          type: ["integer", "null"]
        fileSizeKb:
          type: ["integer", "null"]
        isFeatured:
          type: boolean
        isVisible:
          type: boolean
        isDeleted:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
      required: [id, sectionId, title, board, classLevel, resourceType, isFree, pricePaise, isNew, isFeatured, isVisible, isDeleted, sortOrder, previewImageUrls, createdAt]
...
    RazorpayOrderInput:
      type: object
      properties:
        cardId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
      required: [cardId, customerName, customerPhone]
    RazorpayOrderResult:
      type: object
      properties:
        razorpayOrderId:
          type: string
        orderId:
          type: string
        amountPaise:
          type: integer
        currency:
          type: string
        keyId:
          type: string
        cardTitle:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
      required: [razorpayOrderId, orderId, amountPaise, currency, keyId, cardTitle, customerName, customerPhone]
    VerifyCheckoutInput:
      type: object
      properties:
        razorpayOrderId:
          type: string
        razorpayPaymentId:
          type: string
        razorpaySignature:
          type: string
      required: [razorpayOrderId, razorpayPaymentId, razorpaySignature]
    FreeDownloadInput:
      type: object
      properties:
        cardId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
      required: [cardId, customerName, customerPhone]
    FreeDownloadResult:
      type: object
      properties:
        orderId:
          type: string
        downloadUrl:
          type: string
        filename:
          type: string
      required: [orderId, downloadUrl, filename]
    Order:
      type: object
      properties:
        id:
          type: string
        cardId:
          type: integer
        cardTitle:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        amountPaise:
          type: integer
        status:
          type: string
        razorpayOrderId:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, cardId, cardTitle, customerName, customerPhone, amountPaise, status, createdAt]
...