# lib/api-zod/src/generated/types/freeDownloadInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 2PUC Notes Store API
 * OpenAPI spec version: 0.2.0
 */

export interface FreeDownloadInput {
  cardId: number;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
}