# lib/api-zod/src/generated/types/freeDownloadResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 2PUC Notes Store API
 * OpenAPI spec version: 0.2.0
 */

export interface FreeDownloadResult {
  orderId: string;
  downloadUrl: string;
  filename: string;
}