# lib/api-zod/src/generated/types/razorpayOrderResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 2PUC Notes Store API
 * OpenAPI spec version: 0.2.0
 */

export interface RazorpayOrderResult {
  razorpayOrderId: string;
  orderId: string;
  amountPaise: number;
  currency: string;
  keyId: string;
  cardTitle: string;
  customerName: string;
  customerEmail?: string;
  customerPhone: string;
}