# lib/api-zod/src/generated/types/verifyCheckoutInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 2PUC Notes Store API
 * OpenAPI spec version: 0.2.0
 */

export interface VerifyCheckoutInput {
  razorpayOrderId: string;
  razorpayPaymentId: string;
  razorpaySignature: string;
}